FIX: Make trigger on Signal* return completed Status objects.
